Split CardList into 2 components, adds Card.jsx
